+++ packages/core/src/app/payment/paymentMethod/PaymentMethodTitle.tsx
@@ -5,7 +5,7 @@ import { compact } from 'lodash';
 import React, { type FunctionComponent, memo, type ReactNode } from 'react';
 
 import { BigCommercePaymentsPayLaterBanner } from '@bigcommerce/checkout/bigcommerce-payments-utils'
-import { type CheckoutContextProps, useThemeContext } from '@bigcommerce/checkout/contexts';
+import { type CheckoutContextProps } from '@bigcommerce/checkout/contexts';
 import { withLanguage, type WithLanguageProps } from '@bigcommerce/checkout/locale';
 import { type PaymentFormValues } from '@bigcommerce/checkout/payment-integration-api';
 import { BraintreePaypalCreditBanner, PaypalCommerceCreditBanner } from '@bigcommerce/checkout/paypal-utils';
@@ -395,7 +395,6 @@ const PaymentMethodTitle: FunctionComponent<
 > = ({ cdnBasePath, checkoutSettings, storeCountryCode, onUnhandledError, formik: { values }, instruments, isSelected, language, method, storeCurrency }) => {
     const methodName = getPaymentMethodName(language)(method);
     const { logoUrl, titleText, titleSubText, subtitle } = getPaymentMethodTitle(language, cdnBasePath, checkoutSettings, storeCountryCode, storeCurrency)(method);
-    const { themeV2 } = useThemeContext();
 
     const getSelectedCardType = () => {
         if (!isSelected) {
@@ -494,8 +493,7 @@ const PaymentMethodTitle: FunctionComponent<
                 )} */}
 
                 {titleText && (
-                    <div className={classNames('paymentProviderHeader-name',
-                        { 'sub-header': themeV2 })}
+                    <div className="paymentProviderHeader-name sub-header"
                         data-test="payment-method-name" 
                         style={method.id === 'partially' ? {display: 'block'} : {display: 'contents'}}>
                         {method.id === 'flex'