+++ packages/core/src/app/shipping/PayPalFastlaneShippingAddress.tsx
@@ -118,6 +118,7 @@ export const PayPalFastlaneShippingAddress: FC<PayPalFastlaneShippingAddressProp
                     onAddressSelect={onAddressSelect}
                     onFieldChange={handleFieldChange}
                     onUseNewAddress={props.onUseNewAddress}
+                    validateMaxLength={false}
                     storeCurrencyCode={storeCurrencyCode}
                 />
             )}