+++ packages/stripe-integration/src/stripe-ocs/StripeOCSPaymentMethod.test.tsx
@@ -7,7 +7,10 @@ import {
     type PaymentMethod,
     type WithStripeOCSPaymentInitializeOptions,
 } from '@bigcommerce/checkout-sdk';
-import { createStripeOCSPaymentStrategy } from '@bigcommerce/checkout-sdk/integrations/stripe';
+import {
+    createStripeCSPaymentStrategy,
+    createStripeOCSPaymentStrategy,
+} from '@bigcommerce/checkout-sdk/integrations/stripe';
 import { render } from '@testing-library/react';
 import { Formik } from 'formik';
 import { noop } from 'lodash';
@@ -140,7 +143,7 @@ describe('when using Stripe OCS payment', () => {
         expect(initializePayment).toHaveBeenCalledWith({
             gatewayId,
             methodId,
-            integrations: [createStripeOCSPaymentStrategy],
+            integrations: [createStripeOCSPaymentStrategy, createStripeCSPaymentStrategy],
             [gatewayId]: {
                 containerId: expectedContainerId,
                 layout: defaultAccordionLayout,
@@ -171,7 +174,7 @@ describe('when using Stripe OCS payment', () => {
         expect(initializePayment).toHaveBeenCalledWith({
             gatewayId,
             methodId,
-            integrations: [createStripeOCSPaymentStrategy],
+            integrations: [createStripeOCSPaymentStrategy, createStripeCSPaymentStrategy],
             [gatewayId]: {
                 containerId: expectedContainerId,
                 layout: defaultAccordionLayout,